Blocking request handler
In some situation the order of request/response handling is
important. However, in current implementation the order can be
broken: request/response come from same connection can be handled
in random order because request handler is called in a new thread.
This patch changes that behavior by calling the handler in the same
context, to ensure the ordering.