Write request into codec asynchronously
The c.Go method currently writes request object into channel buffer
synchronously which makes the caller to wait indefinitely in case of
accidental disconnection of server.
Hence this commit makes to write requests asynchronously so that
call object can be returned without any wait and caller can decide
on what to do with it.

Signed-off-by: Periyasamy Palanisamy <pepalani@redhat.com>